requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-multipart==0.0.10
pydub==0.25.1
python-dotenv==1.0.1
